move survey-simple functionality into about module for neatness
