--- static/src/javascripts/projects/commercial/views/survey/survey-simple.html
@@ -1,29 +1,0 @@
-<div class="survey-overlay-simple js-survey-overlay u-h" data-link-name="hosted page about overlay" role="dialog" aria-label="about hosted content">
-    <div class="survey-container">
-        <h3 class="survey-text__header">
-            <%=header%>
-            <% if (showCloseBtn) { %>
-            <div class="survey-close js-survey-close">
-                <button class="site-message__close-btn js-site-message-close" data-link-name="hide about hosted message">
-                    <span class="u-h">Close</span>
-                    <%=crossIcon%>
-                </button>
-            </div>
-            <% } %>
-        </h3>
-        <div class="survey-icon">
-            <%=paidContent%>
-        </div>
-        <div class="survey-text">
-            <p class="survey-text__paragraph">
-                <%=paragraph1%>
-            </p>
-            <p class="survey-text__paragraph">
-                <%=paragraph2%>
-            </p>
-            <p class="survey-text__paragraph">
-                <%=paragraph3%>
-            </p>
-        </div>
-    </div>
-</div>
